feat: renaming prettier package and ignoring from publishing
