+++ .changeset/config.json
@@ -5,7 +5,7 @@
   "changelog": "@changesets/cli/changelog",
   "commit": false,
   "fixed": [],
-  "ignore": ["config-tsconfig", "@acme/docs"],
+  "ignore": ["config-prettier", "config-tsconfig", "@acme/docs"],
   "linked": [],
   "updateInternalDependencies": "patch"
 }
