+++ apps/docs/package.json
@@ -18,8 +18,8 @@
     "@storybook/addon-links": "^6.4.18",
     "@storybook/builder-vite": "^0.2.2",
     "@storybook/react": "^6.4.18",
+    "config-prettier": "0.0.0",
     "config-tsconfig": "0.0.0",
-    "prettier-config-acme": "0.0.0",
     "serve": "^13.0.2",
     "typescript": "^4.5.4",
     "vite": "^3.0.8"
