+++ packages/acme-core/package.json
@@ -24,11 +24,11 @@
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
     "@vitejs/plugin-react": "^2.0.1",
+    "config-prettier": "0.0.0",
     "config-tsconfig": "0.0.0",
     "cypress": "^10.5.0",
     "eslint": "^8.15.0",
     "eslint-config-acme": "0.0.0",
-    "prettier-config-acme": "0.0.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "tsup": "^5.10.1",
