+++ packages/config-prettier/index.js
