+++ packages/config-prettier/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "prettier-config-acme",
+  "name": "config-prettier",
   "version": "0.0.0",
   "main": "index.js",
   "license": "MIT"
