+++ packages/eslint-config-acme/package.json
@@ -6,9 +6,9 @@
   "dependencies": {
     "@typescript-eslint/eslint-plugin": "^5.33.0",
     "@typescript-eslint/parser": "^5.33.0",
+    "config-prettier": "0.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-config-turbo": "latest",
-    "prettier-config-acme": "0.0.0"
+    "eslint-config-turbo": "latest"
   }
 }
