+++ packages/tds-utils/package.json
@@ -31,11 +31,11 @@
     "@types/jest": "28.1.7",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
+    "config-prettier": "0.0.0",
     "config-tsconfig": "0.0.0",
     "eslint": "^8.15.0",
     "eslint-config-acme": "0.0.0",
     "jest": "^28.1.3",
-    "prettier-config-acme": "0.0.0",
     "react": "^18.1.0",
     "react-test-renderer": "^18.1.0",
     "ts-jest": "^28.0.8",
